A very minimal UI for the Analytics frontend. Using knockout.js and jQuery to interact with API via AJAX and GET data to plot with jQplot. Added new functions to IMetric class. Metric name and description so that the Analytics front end will have information about the metrics. Added a function that responds with what parameters each function will need to compute metric data.

+++ Kudu.Core/AnalyticsDataLayer/LogParser.cs
@@ -147,7 +147,7 @@ namespace Kudu.Core.AnalyticsDataLayer
                                     _log.StatusCode = Convert.ToInt16(data[i]);
                                     break;
                                 case W3C_ExtendedConstants.WIN32_STATUS:
-                                    _log.Win32_Status = Convert.ToInt16(data[i]);
+                                    _log.Win32_Status = Convert.ToUInt32(data[i]);
                                     break;
                                 case W3C_ExtendedConstants.COMMENT:
                                     //nothing
